Final version formatter + tests

+++ Zadanie2/Zadanie2/ClassA.cs
@@ -23,7 +23,7 @@ namespace Zadanie2
         {
             Name = info.GetString("name");
             Date = info.GetDateTime("date");
-            Num = (float)info.GetDouble("num");
+            Num = info.GetSingle("num");
             ClassB = (ClassB)info.GetValue("classB", typeof(ClassB));
             ClassC = (ClassC)info.GetValue("classC", typeof(ClassC));
