+++ Zadanie2/Zadanie2/ClassC.cs
@@ -26,7 +26,7 @@ namespace Zadanie2
         {
             Name = info.GetString("name");
             Date = info.GetDateTime("date");
-            Num = (float) info.GetDouble("num");
+            Num = info.GetSingle("num");
             ClassA = (ClassA) info.GetValue("classA", typeof(ClassA));
             ClassB = (ClassB) info.GetValue("classB", typeof(ClassB));
